requirements.txt
fastapi
uvicorn
sqlalchemy
psycop2-binary
pydantic
python-dotenv
apscheduler